+++ test/web_smoke_test.dart
@@ -1,3 +1,6 @@
+@TestOn('browser')
+library;
+
 import 'package:test/test.dart';
 
 void main() {